+++ EduOS.Core/Entities/Attendance/StudentAttendance.cs
@@ -6,7 +6,7 @@ namespace EduOS.Core.Entities.Attendance
 {
     public class StudentAttendance : BaseTenantEntity
     {
-        public int StudentId { get; set; }
+        public long StudentId { get; set; }
         public int ClassId { get; set; }
         public int SectionId { get; set; }
         public DateTime Date { get; set; }
@@ -14,7 +14,7 @@ namespace EduOS.Core.Entities.Attendance
         public TimeSpan? InTime { get; set; }
         public TimeSpan? OutTime { get; set; }
         public string? Remarks { get; set; }
-        public int MarkedBy { get; set; }
+        public long MarkedBy { get; set; }
 
         public virtual Student? Student { get; set; }
         public virtual Class? Class { get; set; }
