+++ EduOS.Core/Interfaces/IRepositories/IStudentAttendanceRepository.cs
@@ -5,11 +5,11 @@ namespace EduOS.Core.Interfaces.IRepositories
     public interface IStudentAttendanceRepository : IGenericRepository<StudentAttendance>
     {
         Task<List<StudentAttendance>> GetByDateAsync(DateTime date, int classId, int sectionId);
-        Task<List<StudentAttendance>> GetByStudentRangeAsync(int studentId, DateTime fromDate, DateTime toDate);
-        Task<StudentAttendance?> GetByStudentAndDateAsync(int studentId, DateTime date);
-        Task<bool> IsAlreadyMarkedAsync(int studentId, DateTime date);
-        Task<int> GetPresentCountAsync(int studentId, DateTime fromDate, DateTime toDate);
-        Task<int> GetAbsentCountAsync(int studentId, DateTime fromDate, DateTime toDate);
-        Task<Dictionary<string, int>> GetMonthlyStatsAsync(int studentId, int month, int year);
+        Task<List<StudentAttendance>> GetByStudentRangeAsync(long studentId, DateTime fromDate, DateTime toDate);
+        Task<StudentAttendance?> GetByStudentAndDateAsync(long studentId, DateTime date);
+        Task<bool> IsAlreadyMarkedAsync(long studentId, DateTime date);
+        Task<int> GetPresentCountAsync(long studentId, DateTime fromDate, DateTime toDate);
+        Task<int> GetAbsentCountAsync(long studentId, DateTime fromDate, DateTime toDate);
+        Task<Dictionary<string, int>> GetMonthlyStatsAsync(long studentId, int month, int year);
     }
 }
